+++ package.json
@@ -55,6 +55,7 @@
     "redux": "^5.0.1",
     "rehype-katex": "^7.0.0",
     "remark-math": "^6.0.0",
+    "sharp": "^0.33.4",
     "styled-components": "^6.1.8",
     "vanilla-tilt": "^1.8.1"
   },
