+++ src/pages/Buttons/top/ScrollToTopButton.jsx
@@ -1,0 +1,37 @@
+import React, { useEffect, useState } from "react";
+import { FaArrowUp } from "react-icons/fa";
+import styles from "./ScrollToTopButton.module.css";
+
+export default function ScrollToTopButton() {
+  const [showButton, setShowButton] = useState(false);
+
+  const scrollToTop = () => {
+    window.scrollTo({
+      top: 0,
+      behavior: "smooth",
+    });
+  };
+
+  const handleScroll = () => {
+    if (window.scrollY > 100) {
+      setShowButton(true);
+    } else {
+      setShowButton(false);
+    }
+  };
+
+  useEffect(() => {
+    window.addEventListener("scroll", handleScroll);
+    return () => {
+      window.removeEventListener("scroll", handleScroll);
+    };
+  }, []);
+
+  return (
+    showButton && (
+      <button onClick={scrollToTop} className={styles.scrollToTopButton}>
+        <FaArrowUp />
+      </button>
+    )
+  );
+}
