+++ src/pages/index.module.css
@@ -51,23 +51,3 @@
 ::-webkit-scrollbar-thumb:hover {
   background: darkgrey;
 }
-.scrollToBottomButton {
-  position: fixed;
-  bottom: 20px;
-  right: 20px;
-  top: auto;
-  background-color: var(--ifm-color-primary);
-  color: #fff;
-  border: none;
-  border-radius: 50%;
-  padding: 10px;
-  box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.1); 
-  cursor: pointer;
-  z-index: 1000;
-}
-
-.scrollToBottomButton:hover {
-  background-color: var(--ifm-color-primary);
-  box-shadow: 0px 6px 8px rgba(0, 0, 0, 0.2); 
-}
-
